pyproject.toml
[tool.poetry]
name = "django-backoffice"
version = "1.0.0"
description = "A set of views, tools and helpers to create a backoffice usingg Django."
authors = ["Marcos Gabarda <hey@marcosgabarda.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.7"
django-model-utils = "^4.0.0"
pytest-cov = "^2.10.0"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
pytest = "^5.4.3"
black = "^19.10b0"
factory_boy = "^2.12.0"
django-test-plus = "^1.4.0"
pytest-django = "^3.9.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
